requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
prisma==0.11.0
pydantic==2.4.2
pydantic-settings==2.0.3
bcrypt==4.0.1
email-validator==2.1.0
requests==2.31.0
twilio==8.10.0

